+++ src/main/kotlin/com/github/andreicostinescu/concepthierarchyplugin/services/ModelService.kt
@@ -4,6 +4,7 @@ import com.github.andreicostinescu.concepthierarchyplugin.settings.RootFileSetti
 import com.github.andreicostinescu.concepthierarchyplugin.utils.JsonIncludeResolver
 import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VfsUtil
@@ -19,13 +20,18 @@ class ModelService(private val project: Project) {
     private var currentModel: Model? = null
 
     fun rebuildModel() {
-        val settings = project.getService(RootFileSettings::class.java)
+        val log = com.intellij.openapi.diagnostic.Logger.getInstance(ModelService::class.java)
+        log.warn("Rebuilding Concept Hierarchy...")
+
+        val settings = project.service<RootFileSettings>()
         val rootPath = settings.getRootPath() ?: return
         val projectBase = project.baseDir ?: project.projectFile?.parent
         val abs =
             if (projectBase != null) VfsUtil.findRelativeFile(rootPath, projectBase) else LocalFileSystem.getInstance()
                 .findFileByPath(rootPath)
         val psiRoot = abs?.let { PsiManager.getInstance(project).findFile(it) as? JsonFile } ?: return
+
+        log.warn("Reading Concept Hierarchy files...")
 
         ReadAction.run<RuntimeException> {
             val includes = JsonIncludeResolver.findIncludes(psiRoot)