+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 ### Added <!-- for new features -->
+- Read the Concept Hierarchy files upon startup, changing of the root file, applying of the settings, and modifying any of the file pertaining to the current concept hierarchy.
 - Ctrl + B with the cursor on an external file goes to included external file (defined under the "external", "data", and "header" keywords)
 - Added an editor action to mark a `.json` file as the root Concept Hierarchy definition file
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)